+++ changelog.md
@@ -9,7 +9,10 @@ This is what we are currently working on:
 - Add a fallback background color to the modifier bucket to fix the transparent MB bug.
 - System setting to check for equipped status when determining weight (players can use equipped flag to indicate a 'dropped' item)
 - OtF now supports [Sk:Skillname] or [Sp:Spellname]
-- Added support for the Drag Ruler module (displaying move, sprint and fly distances)
+- Added support for the Drag Ruler module to display move, sprint and fly distances. (but you also need to load the Hex Token Size Support module, due to an error with Drag Ruler)
+- changed parameter order of /rem [otf] optionalusernames to match /sendmb
+- Holding down CTRL (CMD on Macs) will make a roll BLIND
+- System setting "SHIFT Click does a Blind roll for players, and a Self roll for GMs"
 
 ## History